+++ CHANGELOG.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2023-06-??
+
+### Added
+
+* Foundry 11 support
+
+### Removed
+
+* Foundry 10 support
+
+### Changed
+
+* Show unrest penalty in sidebar
+* Group compendia in new compendia folders
+* Move Settlement config into separate dialog
+
 ## [0.0.35] - 2023-04-27
 
 ### Fixed